feat: introduction of a logical unit
it's purpose is to surround elements that logically belong together.

---

Yes, i hear you, a simple div could do the same, this basically is a
simple div that you can give some class.
And you would be fully right! But this element has a semantic benefit,
by declaring it as a logical element, we also read it is as such. The
intention of this component is not to replace the infamous `div`.
This component exists to group things, and create a better understanding
of which elements really belong together and aren't there for
spacing/beautifying stuff necessarily :), welcome to my ted talk.